Make the backup on update optional
Set this as disabled by default

+++ docker/overlay/init.d/start/30-download-sentry.sh
@@ -5,7 +5,7 @@
 # File Created: Monday, 21st October 2024 11:14:56 am
 # Author: Josh5 (jsunnex@gmail.com)
 # -----
-# Last Modified: Monday, 21st October 2024 10:52:41 pm
+# Last Modified: Friday, 29th November 2024 12:51:13 am
 # Modified By: Josh5 (jsunnex@gmail.com)
 ###
 
@@ -52,17 +52,19 @@ if [ "${PREVIOUS_SENTRY_VERSION:-}" != "${SENTRY_VERSION}" ]; then
     echo "  - Pulling down any existing Sentry stack services..."
     ${docker_compose_cmd:?} down --remove-orphans
 
-    echo "  - Backup Docker volumes..."
-    volumes="sentry-data sentry-postgres sentry-redis sentry-zookeeper sentry-kafka sentry-clickhouse sentry-symbolicator"
-    mkdir -p "${SENTRY_DATA_PATH}"/update_backups/self_hosted-v${PREVIOUS_SENTRY_VERSION:-UNKNOWN}-${date_string}/volumes
-    for volume_name in ${volumes}; do
-        if ${docker_cmd:?} volume inspect "${volume_name}" >/dev/null 2>&1; then
-            echo "    - Backing up ${volume_name}..."
-            ${docker_cmd:?} run --rm -v "${volume_name}":/data -v "${SENTRY_DATA_PATH}"/update_backups/self_hosted-v${PREVIOUS_SENTRY_VERSION:-UNKNOWN}-${date_string}/volumes:/backup ubuntu tar -zcvf /backup/${volume_name}.tar /data
-        else
-            echo "  Volume ${volume_name} not found, skipping backup."
-        fi
-    done
+    if [ "${BACKUP_VOLUMES_ON_UPDATE:-}" = "true" ]; then
+        echo "  - Backup Docker volumes..."
+        volumes="sentry-data sentry-postgres sentry-redis sentry-zookeeper sentry-kafka sentry-clickhouse sentry-symbolicator"
+        mkdir -p "${SENTRY_DATA_PATH}"/update_backups/self_hosted-v${PREVIOUS_SENTRY_VERSION:-UNKNOWN}-${date_string}/volumes
+        for volume_name in ${volumes}; do
+            if ${docker_cmd:?} volume inspect "${volume_name}" >/dev/null 2>&1; then
+                echo "    - Backing up ${volume_name}..."
+                ${docker_cmd:?} run --rm -v "${volume_name}":/data -v "${SENTRY_DATA_PATH}"/update_backups/self_hosted-v${PREVIOUS_SENTRY_VERSION:-UNKNOWN}-${date_string}/volumes:/backup ubuntu tar -zcvf /backup/${volume_name}.tar /data
+            else
+                echo "  Volume ${volume_name} not found, skipping backup."
+            fi
+        done
+    fi
 
     echo "  - Install volumes restore script..."
     echo "${RESTORE_VOLUMES_SCRIPT:?}" >"${SENTRY_DATA_PATH}"/update_backups/self_hosted-v${PREVIOUS_SENTRY_VERSION:-UNKNOWN}-${date_string}/volumes/restore-volumes.sh
